Aggregate legend query and query builder buttons
