+++ .changeset/silent-moose-itch.md
@@ -1,0 +1,8 @@
+---
+'@finos/legend-extension-dsl-data-space-studio': patch
+'@finos/legend-application-query-bootstrap': patch
+'@finos/legend-extension-dsl-data-space': patch
+'@finos/legend-extension-dsl-service': patch
+'@finos/legend-application-studio': patch
+'@finos/legend-query-builder': patch
+---
