+++ .changeset/soft-bugs-argue.md
@@ -1,0 +1,5 @@
+---
+'@finos/legend-application-query': patch
+---
+
+Aggregate Legend Query Application header actions with Query Builder Header actions
