+++ packages/legend-application-query/src/components/__tests__/QueryEditorQueryActions.test.tsx
@@ -48,7 +48,6 @@ import {
   TEST_DATA__modelCoverageAnalysisResult,
   TEST_DATA__simpleProjectionQuery,
 } from './TEST_DATA__QueryBuilder_ResultStateTest.js';
-import { QUERY_EDITOR_TEST_ID } from '../../__lib__/LegendQueryTesting.js';
 
 test(
   integrationTest(
@@ -99,10 +98,10 @@ test(
     });
 
     await waitFor(() =>
-      renderResult.getByTestId(QUERY_EDITOR_TEST_ID.QUERY_EDITOR_ACTIONS),
+      renderResult.getByTestId(QUERY_BUILDER_TEST_ID.QUERY_BUILDER_ACTIONS),
     );
     const queryActionsPanel = renderResult.getByTestId(
-      QUERY_EDITOR_TEST_ID.QUERY_EDITOR_ACTIONS,
+      QUERY_BUILDER_TEST_ID.QUERY_BUILDER_ACTIONS,
     );
 
     expect(
@@ -209,10 +208,10 @@ test(
     });
 
     await waitFor(() =>
-      renderResult.getByTestId(QUERY_EDITOR_TEST_ID.QUERY_EDITOR_ACTIONS),
+      renderResult.getByTestId(QUERY_BUILDER_TEST_ID.QUERY_BUILDER_ACTIONS),
     );
     const queryActionsPanel = renderResult.getByTestId(
-      QUERY_EDITOR_TEST_ID.QUERY_EDITOR_ACTIONS,
+      QUERY_BUILDER_TEST_ID.QUERY_BUILDER_ACTIONS,
     );
 
     expect(
@@ -295,7 +294,7 @@ test(
 
     // Verify action buttons are disabled properly and error is shown
     const queryActionsPanel = await renderResult.findByTestId(
-      QUERY_EDITOR_TEST_ID.QUERY_EDITOR_ACTIONS,
+      QUERY_BUILDER_TEST_ID.QUERY_BUILDER_ACTIONS,
     );
     expect(
       getByRole(queryActionsPanel, 'button', {
@@ -389,7 +388,7 @@ test(
 
     // Verify action buttons are disabled properly and no error is shown
     const queryActionsPanel = await renderResult.findByTestId(
-      QUERY_EDITOR_TEST_ID.QUERY_EDITOR_ACTIONS,
+      QUERY_BUILDER_TEST_ID.QUERY_BUILDER_ACTIONS,
     );
     expect(
       getByRole(queryActionsPanel, 'button', {
