+++ packages/legend-application-query/src/stores/LegendQueryApplicationPlugin.tsx
@@ -16,12 +16,14 @@
 
 import { LegendApplicationPlugin } from '@finos/legend-application';
 import type { Query } from '@finos/legend-graph';
-import type { QueryBuilderState } from '@finos/legend-query-builder';
-import type React from 'react';
-import type { LegendQueryPluginManager } from '../application/LegendQueryPluginManager.js';
 import type {
-  ExistingQueryEditorStore,
-  QueryEditorStore,
+  QueryBuilderState,
+  QueryBuilder_LegendApplicationPlugin_Extension,
+} from '@finos/legend-query-builder';
+import type { LegendQueryPluginManager } from '../application/LegendQueryPluginManager.js';
+import {
+  type ExistingQueryEditorStore,
+  type QueryEditorStore,
 } from './QueryEditorStore.js';
 import type { QuerySetupLandingPageStore } from './QuerySetupStore.js';
 
@@ -77,7 +79,10 @@ export type QueryEditorHelpMenuActionConfiguration = {
   icon?: React.ReactNode;
 };
 
-export abstract class LegendQueryApplicationPlugin extends LegendApplicationPlugin {
+export class LegendQueryApplicationPlugin
+  extends LegendApplicationPlugin
+  implements QueryBuilder_LegendApplicationPlugin_Extension
+{
   /**
    * This helps to better type-check for this empty abtract type
    * See https://github.com/finos/legend-studio/blob/master/docs/technical/typescript-usage.md#understand-typescript-structual-type-system
