+++ packages/legend-application-query/style/_query-editor.scss
@@ -22,7 +22,7 @@
   width: 100%;
   height: 100%;
 
-  &__header {
+  &__logo-header {
     @include flexVCenter;
 
     height: 3.4rem;
@@ -33,6 +33,7 @@
       height: 3.4rem;
       width: 5rem;
       border-right: 0.1rem solid var(--color-dark-grey-200);
+      background: var(--color-dark-grey-100);
     }
 
     &__menu-item {
@@ -40,12 +41,26 @@
 
       height: 3.4rem;
       width: 100%;
-      color: var(--color-dark-grey-400);
+      color: var(--color-dark-grey-500);
 
       svg {
         font-size: 2.3rem;
       }
     }
+
+    &__name {
+      color: var(--color-dark-grey-500);
+      font-size: 2.5rem;
+      font-weight: 700;
+      margin-left: 1rem;
+    }
+  }
+
+  &__header {
+    @include flexVCenter;
+
+    height: 3.4rem;
+    padding: 0.2rem;
 
     &__content {
       @include flexVCenter;
@@ -71,6 +86,9 @@
         margin: 0 1rem;
         margin-right: 10%;
         cursor: pointer;
+        color: var(--color-dark-grey-400);
+        font-size: 1.5rem;
+        font-weight: 500;
       }
     }
