+++ packages/legend-extension-dsl-data-space/src/stores/query-builder/DataSpaceQueryBuilderState.ts
@@ -17,9 +17,10 @@
 import { type GenericLegendApplicationStore } from '@finos/legend-application';
 import {
   type QueryBuilderConfig,
-  QueryBuilderState,
   type QuerySDLC,
   type QueryBuilderWorkflowState,
+  type QueryBuilderActionConfig,
+  QueryBuilderState,
 } from '@finos/legend-query-builder';
 import {
   type GraphManagerState,
@@ -155,6 +156,7 @@ export class DataSpaceQueryBuilderState extends QueryBuilderState {
     graphManagerState: GraphManagerState,
     depotServerClient: DepotServerClient,
     workflow: QueryBuilderWorkflowState,
+    actionConfig: QueryBuilderActionConfig,
     dataSpace: DataSpace,
     executionContext: DataSpaceExecutionContext,
     onDataSpaceChange: (val: DataSpaceInfo) => void,
@@ -195,6 +197,7 @@ export class DataSpaceQueryBuilderState extends QueryBuilderState {
     this.onClassChange = onClassChange;
     this.isAdvancedDataSpaceSearchEnabled = isAdvancedDataSpaceSearchEnabled;
     this.dataSpaceAnalysisResult = dataSpaceAnalysisResult;
+    this.workflowState.updateActionConfig(actionConfig);
   }
 
   override get sideBarClassName(): string | undefined {
