+++ packages/legend-query-builder/src/__lib__/QueryBuilderTesting.ts
@@ -18,6 +18,7 @@ export enum QUERY_BUILDER_TEST_ID {
   LAMBDA_EDITOR__EDITOR_INPUT = 'lambda-editor__editor__input',
 
   QUERY_BUILDER = 'query__builder',
+  QUERY_BUILDER_ACTIONS = 'query__builder__actions',
   QUERY_BUILDER_SETUP = 'query__builder__setup',
   QUERY_BUILDER_FETCH_STRUCTURE = 'query__builder__fetch__structure',
   QUERY_BUILDER_TDS_PROJECTION = 'query__builder__tds__projection',
