+++ packages/legend-query-builder/src/stores/QueryBuilder_LegendApplicationPlugin_Extension.ts
@@ -75,6 +75,14 @@ export type TemplateQueryPanelContentRenderer = (
   queryBuilderState: QueryBuilderState,
 ) => React.ReactNode;
 
+export type QueryBuilderHeaderActionConfiguration = {
+  key: string;
+  category: number;
+  renderer: (
+    queryBuilderState: QueryBuilderState,
+  ) => React.ReactNode | undefined;
+};
+
 export interface QueryBuilder_LegendApplicationPlugin_Extension
   extends LegendApplicationPlugin {
   /**
@@ -118,4 +126,14 @@ export interface QueryBuilder_LegendApplicationPlugin_Extension
    * Get the list of template query panel content render
    */
   getExtraTemplateQueryPanelContentRenderer?(): TemplateQueryPanelContentRenderer[];
+
+  /**
+   * Get the list of action configurations
+   */
+  getExtraQueryBuilderHeaderActionConfigurations?(): QueryBuilderHeaderActionConfiguration[];
+
+  /**
+   * Get the list of action configurations
+   */
+  getExtraQueryBuilderHeaderTitleConfigurations?(): QueryBuilderHeaderActionConfiguration[];
 }
