+++ packages/legend-query-builder/src/stores/workflows/ServiceQueryBuilderState.ts
@@ -34,7 +34,10 @@ import { action, makeObservable, observable } from 'mobx';
 import { renderServiceQueryBuilderSetupPanelContent } from '../../components/workflows/ServiceQueryBuilder.js';
 import { QueryBuilderState } from '../QueryBuilderState.js';
 import type { QueryBuilderConfig } from '../../graph-manager/QueryBuilderConfig.js';
-import type { QueryBuilderWorkflowState } from '../query-workflow/QueryBuilderWorkFlowState.js';
+import type {
+  QueryBuilderActionConfig,
+  QueryBuilderWorkflowState,
+} from '../query-workflow/QueryBuilderWorkFlowState.js';
 
 export type ServiceExecutionContext = {
   key: string;
@@ -60,6 +63,7 @@ export class ServiceQueryBuilderState extends QueryBuilderState {
     applicationStore: GenericLegendApplicationStore,
     graphManagerState: GraphManagerState,
     workflowState: QueryBuilderWorkflowState,
+    actionConfig: QueryBuilderActionConfig,
     service: Service,
     usableServices: Service[] | undefined,
     executionContextKey?: string | undefined,
@@ -87,6 +91,7 @@ export class ServiceQueryBuilderState extends QueryBuilderState {
     this.usableServices = usableServices;
     this.onServiceChange = onServiceChange;
     this.onExecutionContextChange = onExecutionContextChange;
+    this.workflowState.updateActionConfig(actionConfig);
 
     if (service.execution instanceof PureSingleExecution) {
       this.executionContextState.mapping = service.execution.mapping?.value;
